pyproject.toml
[project]
name = "paper-sanitizer"
version = "0.1.0"
description = "CLI that strips prompt-injection triggers from academic papers"
authors = [{ name = "Thordur Arnason", email = "thordur@gmail.com" }]
requires-python = ">=3.9"
dependencies = ["pdfminer.six>=20221105"]

[project.scripts]
sanitize = "sanitizer.cli:main" 
